Restructure into using MVC design pattern
