+++ main.py
@@ -2,16 +2,14 @@ import sys
 from PySide2.QtWidgets import QApplication, QMainWindow
 from PySide2.QtGui import Qt, QIcon
 
-from application.application import Application
-
-from main_window.main_window import MainWindow
-from toolbar.toolbar import Toolbar
-from status_bar.status_bar import StatusBar
-from menubar.menubar import Menubar
-from dock.dock import Dock
-
-from central_widget.central_widget import CentralWidget
-from workspace.workspace import Workspace
+from view.application.application import Application
+from view.main_window.main_window import MainWindow
+from view.toolbar.toolbar import Toolbar
+from view.status_bar.status_bar import StatusBar
+from view.menubar.menubar import Menubar
+from view.dock.dock import Dock
+from view.central_widget.central_widget import CentralWidget
+from view.workspace.workspace import Workspace
 
 if __name__ == "__main__":
     app = Application(sys.argv)
@@ -26,8 +24,8 @@ if __name__ == "__main__":
     window.addToolBar(Qt.LeftToolBarArea, toolbar)
 
     central_widget = CentralWidget(window)
-    window.setCentralWidget(central_widget) 
-    
+    window.setCentralWidget(central_widget)
+
     status_bar = StatusBar(window)
     window.setStatusBar(status_bar)
