--- main_window/main_window.py
@@ -1,8 +1,0 @@
-from PySide2.QtWidgets import QMainWindow
-
-
-class MainWindow(QMainWindow):
-    def __init__(self):
-        QMainWindow.__init__(self)
-        self.setWindowTitle("Zephyrus - Information Handler")
-
