+++ model/meta/firms_metadata.json
@@ -1,0 +1,8 @@
+{
+  "search key": "city",
+  "columns": ["firm name", "employee number", "profit"],
+  "handler type": "serial",
+  "linked file": "countries.json",
+  "additional tab name": "countries",
+  "subtable columns": ["country", "city"]
+}
