+++ model/storage/countries.json
@@ -1,0 +1,14 @@
+[
+  {
+    "country": "Serbia",
+    "city": "Novi Sad"
+  },
+  {
+    "country": "Germany",
+    "city": "Berlin"
+  },
+  {
+    "country": "Austria",
+    "city": "Wien"
+  }
+]
