+++ model/storage/days.json
@@ -1,0 +1,30 @@
+[
+    {
+        "day_id": 0,
+        "Day of the week": "Monday"
+    },
+    {
+        "day_id": 1,
+        "Day of the week": "Tuesday"
+    },
+    {
+        "day_id": 2,
+        "Day of the week": "Wednesday"
+    },
+    {
+        "day_id": 3,
+        "Day of the week": "Thursday"
+    },
+    {
+        "day_id": 4,
+        "Day of the week": "Friday"
+    },
+    {
+        "day_id": 5,
+        "Day of the week": "Saturday"
+    },
+    {
+        "day_id": 6,
+        "Day of the week": "Sunday"
+    }
+]
