+++ model/storage/tasks.json
@@ -1,0 +1,47 @@
+[
+    {
+        "day_id": 0,
+        "Name": "Buy groceries",
+        "Time": "13:50"
+    },
+    {
+        "day_id": 1,
+        "Name": "Prepare lunch",
+        "Time": "12:35"
+    },
+    {
+        "day_id": 1,
+        "Name": "Walk the dog",
+        "Time": "11:00"
+    },
+    {
+        "day_id": 2,
+        "Name": "Exercise",
+        "Time": "20:00"
+    },
+    {
+        "day_id": 2,
+        "Name": "Go shopping",
+        "Time": "13:50"
+    },
+    {
+        "day_id": 3,
+        "Name": "Feed the rabbit",
+        "Time": "12:35"
+    },
+    {
+        "day_id": 5,
+        "Name": "curee",
+        "Time": "20:00"
+    },
+    {
+        "day_id": 5,
+        "Name": "Go to war",
+        "Time": "11:00"
+    },
+    {
+        "day_id": 5,
+        "Name": "Cure COVID-19",
+        "Time": "23:00"
+    }
+]
