--- storage/countries.json
@@ -1,14 +1,0 @@
-[
-    {
-        "country": "Serbia",
-        "city": "Novi Sad"
-    },
-    {
-        "country": "Germany",
-        "city": "Berlin"
-    },
-    {
-        "country": "Austria",
-        "city": "Wien"
-    }
-]
