--- storage/countries_metadata.json
@@ -1,9 +1,0 @@
-{
-    "search key": "city",
-    "columns": ["country", "city"],
-    "handler type": "serial",
-    "additional tab name": "firms",
-    "linked file": "firms.json",
-    "subtable columns": ["firm name", "employee number", "profit"]
-
-}
