--- style_file.qss
@@ -1,92 +1,0 @@
-/* QMainWindow {
-    background-color: #212121;
-}
-
-QMenuBar {
-    background-color: #1a1a1a;
-    color: #7c7c7c;
-    spacing: 3px;
-}
-
-QMenuBar::item {
-    padding: 5px 4px;
-    background: transparent;
-    border-radius: 0;
-} */
-
-/* QMenuBar::item:selected { 
-    background: #404040;
-    color: white;
-} */
-
-/* QMenuBar::item:pressed {
-} */
-
-/* ovo ne radi iz nekog razloga */
-/* QMenuBar::item#menu_logo:hover,
-QMenuBar::item#menu_logo:selected, 
-QMenuBar::item#menu_logo:pressed {
-    background: #1a1a1a;
-} */
-
-/* QToolBar {
-    background-color: #1a1a1a;
-    min-width: 50px;
-    spacing: 3px;
-}
-
-QToolButton {
-    margin-top: 5px;
-    width: 3em;
-    border-radius: 0;
-}
-
-QToolButton:hover {
-    background-color: rgba(64,64,64, 0.3);
-    min-width: 50px;
-    border-left: 3px solid #eeffff;
-}
-
-QToolButton:pressed {
-    background-color: #212121;
-}
-
-QStatusBar {
-    color: #7b7b7b;
-    background-color: #1a1a1a;
-} */
-
-/* QDockWidget {
-    color: #828282;
-    background-color: #1a1a1a;
-}
-
-QTreeView {
-    color: #616161;
-    background-color: #1e1e1e;
-}
-
-QHeaderView {
-    color: #616161;
-    background-color: #1e1e1e;
-}
-
-QScrollBar:horizontal {
-    color: black;
-    background-color: #424242;
-    border: none;
-}
-
-QScrollBar:vertical {
-    color: white;
-    background-color: #424242;
-    border: none;
-}
-QTabBar::tab {
-    background-color: #1a1a1a;
-    color: #7c7c7c;
-}
-QTabBar::tab:selected {
-    background-color: #212121;
-    color: white
-}  */
