+++ view/status_bar/status_bar.py
